Add back composability support
Prior to v3, the attributes matcher was defined using the RSpec DSL
which meant that the resulting matcher(s) was a subclass of RSpec's
built-in matcher hierarchy.

This meant that the matchers were chainable, composable, diffable,
and recursive.

The v3 base matcher does not inherit from the core RSpec matchers,
and thus breaks the interfaces that RSpec matchers support.

This changes adds `RSpec::Matchers::Composable` to the base matcher
to add back composability support

Fix #107