Configure RuboCop for RSpec Matcher protocol
Allow the use of `does_not_match?` in the `Naming/PredicateName` cop,
as it aligns with the RSpec Matcher Protocol. This method is part of
RSpec's internal API for defining custom matchers.

This change addresses the stricter predicate naming rules introduced in
RuboCop (rubocop/rubocop#12358), which discourage auxiliary verbs like
`does` in predicate methods.

The exception is necessary here to maintain compatibility with
RSpec's conventions.

Ref:
- https://www.rubydoc.info/github/rspec/rspec-expectations/RSpec/Matchers/MatcherProtocol#does_not_match%3F-instance_method